Multiple_Linear_Regression
There is a dataset regarding sales figures based on TV, radio, and newspaper advertisements. Using this data, the impact of a specific investment on sales figures is being analyzed through machine learning methods. Specifically, the aim is to determine how much sales figures increase with individual investments in TV, radio, and newspapers, and to identify which advertising method is more successful. This analysis will help us understand which advertising channel is more effective and how to allocate our future advertising budget most efficiently.